+++ java_segway/src/java_segway/StartUp.java
@@ -4,7 +4,7 @@ package java_segway;
 public class StartUp {
 	
 	public static void main(String[] args){
-		double regPeriod = 0.02; //s
+		float regPeriod = 0.02f; //s
 //		RegulAndIO reg = new RegulAndIO(regPeriod);
 		FeedbackController reg = new FeedbackController(regPeriod);
 		OpCom opCom = new OpCom(null, reg);
